[CSL-1180] Sketch Delegation.Cede, class/instances/logic
