+++ src/Pos/Delegation/Cede.hs
@@ -1,0 +1,3 @@
+-- | Re-exports of everything from 'Pos.Delegation.Cede*'.
+
+{-# OPTIONS_GHC -F -pgmF autoexporter #-}
